Drop the minifyCss transform and go back to a Css#minify method
Async, though, because of the new cssnano API
