+++ lib/assets/Css.js
@@ -7,6 +7,7 @@ const perfectionist = require('perfectionist');
 const errors = require('../errors');
 const Text = require('./Text');
 const propertyWithImageUrlRegExp = /^(?:content|_?cursor|_?background(?:-image)?|(?:-[a-z]+-)?(?:border-image(?:-source)?|mask|mask-image|mask-image-source|mask-box-image|mask-box-image-source)|(?:\+|-webkit-)?filter)$/;
+const cssnano = require('cssnano');
 
 function extractEncodingFromText(text) {
   const matchCharset = text.match(/@charset\s*(['"])\s*([\w-]+)\s*\1/im);
@@ -206,6 +207,33 @@ class Css extends Text {
     /*eslint-enable*/
     this.markDirty();
     return this;
+  }
+
+  async minify() {
+    try {
+      const result = await cssnano.process(
+        this.parseTree,
+        { map: { annotation: false } }, // Tell postcss to not remove the sourceMappingURL comment
+        {
+          preset: [
+            'default',
+            {
+              svgo: false,
+              discardComments: {
+                remove(comment) {
+                  return !/@preserve|@license|[@#]\s*sourceURL|[#@]\s*sourceMappingURL|^!/.test(
+                    comment
+                  );
+                }
+              }
+            }
+          ]
+        }
+      );
+      this.parseTree = result.root;
+    } catch (err) {
+      this.assetGraph.warn(err);
+    }
   }
 
   findOutgoingRelationsInParseTree() {
