+++ lib/assets/Text.js
@@ -93,8 +93,8 @@ class Text extends Asset {
   set encoding(encoding) {
     if (encoding !== this.encoding) {
       /*eslint-disable*/
-            const text = this.text; // Make sure this._text exists so the rawSrc is decoded before the original encoding is thrown away
-            /*eslint-enable*/
+      const text = this.text; // Make sure this._text exists so the rawSrc is decoded before the original encoding is thrown away
+      /*eslint-enable*/
       this._rawSrc = undefined;
       this._encoding = encoding;
       this.markDirty();
