--- lib/transforms/minifyCss.js
@@ -1,39 +1,0 @@
-const cssnano = require('cssnano');
-
-const defaultQueryObj = { type: 'Css', isLoaded: true };
-
-module.exports = queryObj => {
-  return async assetGraph => {
-    if (queryObj) {
-      queryObj = { $and: [queryObj, defaultQueryObj] };
-    } else {
-      queryObj = defaultQueryObj;
-    }
-    for (const cssAsset of assetGraph.findAssets(queryObj)) {
-      try {
-        const result = await cssnano.process(
-          cssAsset.parseTree,
-          { map: { annotation: false } }, // Tell postcss to not remove the sourceMappingURL comment
-          {
-            preset: [
-              'default',
-              {
-                svgo: false,
-                discardComments: {
-                  remove(comment) {
-                    return !/@preserve|@license|[@#]\s*sourceURL|[#@]\s*sourceMappingURL|^!/.test(
-                      comment
-                    );
-                  }
-                }
-              }
-            ]
-          }
-        );
-        cssAsset.parseTree = result.root;
-      } catch (err) {
-        assetGraph.warn(err);
-      }
-    }
-  };
-};
