+++ lib/transforms/subsetFonts.js
@@ -542,6 +542,11 @@ module.exports = ({
       ]
     };
 
+    // Keep track of the injected CSS assets that should eventually be minified
+    // Minifying them along the way currently doesn't work because some of the
+    // manipulation is sensitive to the exact text contents. We should fix that.
+    const subsetFontsToBeMinified = new Set();
+
     for (const htmlAsset of htmlAssets) {
       const accumulatedFontFaceDeclarations = [];
 
@@ -761,10 +766,10 @@ module.exports = ({
         cssAsset = assetGraph.addAsset({
           type: 'Css',
           url: subsetUrl + 'subfontTemp.css',
-          text: subsetCssText,
-          _subsetFontsToBeMinified: true
+          text: subsetCssText
         });
 
+        subsetFontsToBeMinified.add(cssAsset);
         cssAssetMap[subsetCssText] = cssAsset;
 
         if (!inlineSubsets) {
@@ -1008,11 +1013,8 @@ module.exports = ({
     }
 
     // This is a bit awkward now, but if it's done sooner, it breaks the CSS source regexping:
-    for (const cssAsset of assetGraph.findAssets({
-      _subsetFontsToBeMinified: true
-    })) {
-      await assetGraph.minifyCss(cssAsset);
-      delete cssAsset._subsetFontsToBeMinified;
+    for (const cssAsset of subsetFontsToBeMinified) {
+      await cssAsset.minify();
     }
 
     // Hand out some useful info about the detected subsets:
