+++ testdata/assets/Css/minify/existingSourceMap/foo.css
